requirements.txt
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
aiosqlite==0.17.0
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
Babel==2.9.1
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
environs==9.5.0
fastapi==0.95.0
frozenlist==1.3.3
h11==0.14.0
idna==3.4
iso8601==1.1.0
magic-filter==1.0.9
marshmallow==3.19.0
multidict==6.0.4
packaging==23.0
pydantic==1.10.7
pypika-tortoise==0.1.6
python-dotenv==1.0.0
pytz==2022.7.1
sniffio==1.3.0
starlette==0.26.1
tortoise-orm==0.19.3
typing_extensions==4.5.0
uvicorn==0.21.1
yarl==1.8.2

oauth2client~=4.1.3
apiclient~=1.0.4
httplib2~=0.22.0